requirements.txt
preprocessor
matplotlib
pandas
nltk
alphabet-detector
tweepy
python-bidi
arabic_reshaper
wordcloud
cltk
scikit-learn
elasticsearch